requirements.txt
google-genai
faiss-cpu
numpy
pandas
streamlit
pydantic
python-dotenv
